feat: setup bundling of js files with webpack
Currently, only JS files are bundled. To test, run the following in the
CLI:

npx webpack --mode development ./src/index.js --output ./dist/main

Webpack will bundle the index.js file  (and any dependency) in the
src directory to one file main.js.. and provide the bundling details on
the terminal.

To view the result in a browser, you can create an html page in the /dist
linking main.js as a script file. In the console, we should have:
- Logging from index.js
- Logging from a dependency file