feat: setup bundling of stylesheet
CSS Loader is used, in the src directory, to extract the stylesheet
into the javascript file while Mini CSS Extract Plugin extracts the
stylesheet from the javascript file into it respective stylesheet in the
distribution directory.

To test,
- Run the development build script
- In the dist directory, create a new html file with a text node and link
the generated stylesheet in the head tag
- View the html file in a browser, if the background is black and text
color is white then it works!