fix(test): access dom nodes only if available
With the last test, commit 55bef2f, an invalid html is generated as
htmlwebpack plugin doesn't understand pug files yet. This result to
invalid dom nodes and errors were threw, execution was stopped.

This commit fixes the issue of accessing the nodes only if they are
available in the first place so that all tests can run smoothly